BOLETIN11 ReproductorDeAudio

+++ Boletin11/src/EjercicioClase2/mainReproductorAudio.java
@@ -2,7 +2,7 @@ package EjercicioClase2;
 
 public class mainReproductorAudio {
     public static void main(String[] args) {
-        ReproductorDeAudio reproductorDeAudio = new ReproductorDeAudio("Beats","Primero",3);
+        ReproductorDeAudio reproductorDeAudio = new ReproductorDeAudio("Beats","Primero", (byte) 3);
         System.out.println(reproductorDeAudio);
     }
 }
